Chapter04/app_engine/requirements.txt
Flask==2.2.2
google-cloud-storage==2.5.0
google-cloud-logging==3.2.4